Format href using the react-router formatPattern method
This is a *BREAKING* change as shown by the required change to the tests.
Required parameters that are not present will now throw an invariant exception rather than substituting the key in.
The old behavior could be restored by duplicating the formatPattern code and modifying it appropriately but this would have the downside of allow them to become out of sync again.